requirements.txt
Flask==3.1.1
Flask-Limiter==3.12
gunicorn==21.2.0
Werkzeug==3.1.3
python-dotenv==1.0.0
redis==5.0.1
psutil==5.9.6
requests==2.31.0
# Scientific packages for 3D conversion - Python 3.13 compatible versions
numpy>=1.26.0
scipy>=1.11.0
networkx>=3.2
pycollada>=0.9.1
# Try a lighter version of trimesh or alternative
trimesh>=3.20.0 